Enhance social images

+++ app/Jobs/GenerateSocialImagesJob.php
@@ -42,41 +42,45 @@ class GenerateSocialImagesJob implements ShouldQueue
             $container = AssetContainer::find('social_images');
             $disk = $container->disk();
 
-            // Delete any old images/meta remaining.
-            collect([
-                $item->get('og_image'),
-                $item->get('twitter_image'),
-            ])
-            ->filter()
-            ->each(function ($image) use ($container) {
-                if($container->asset($image)->exists()){
-                    $container->asset($image)->delete();
-                }
-            });
+            // Skip if project type != list.
+            if ($item->get('project_type') !== 'list') {
 
-            // Prepare.
-            $id = $item->id();
-            $title = Str::of($item->get('title'))->slug('-');
-            $absolute_url = $item->site()->absoluteUrl();
-            $unique = time();
+                // Delete any old images/meta remaining.
+                collect([
+                    $item->get('og_image'),
+                    $item->get('twitter_image'),
+                ])
+                ->filter()
+                ->each(function ($image) use ($container) {
+                    if($container->asset($image)->exists()){
+                        $container->asset($image)->delete();
+                    }
+                });
 
-            // Generate, save and set default og image/meta.
-            $file = "{$title}-og-{$unique}.png";
-            $image = Browsershot::url("{$absolute_url}/social-images/{$id}")
-                ->windowSize(1200, 630)
-                ->select('#og')
-                ->save($disk->path($file));
-            $container->makeAsset($file)->save();
-            $item->set('og_image', $file)->save();
+                // Prepare.
+                $id = $item->id();
+                $title = Str::of($item->get('title'))->slug('-');
+                $absolute_url = $item->site()->absoluteUrl();
+                $unique = time();
 
-            // Generate, save and set default twitter image/meta.
-            $file = "{$title}-twitter-{$unique}.png";
-            $image = Browsershot::url("{$absolute_url}/social-images/{$id}")
-                ->windowSize(1200, 600)
-                ->select('#twitter')
-                ->save($disk->path($file));
-            $container->makeAsset($file)->save();
-            $item->set('twitter_image', $file)->save();
+                // Generate, save and set default og image/meta.
+                $file = "{$title}-og-{$unique}.png";
+                $image = Browsershot::url("{$absolute_url}/social-images/{$id}")
+                    ->windowSize(1200, 630)
+                    ->select('#og')
+                    ->save($disk->path($file));
+                $container->makeAsset($file)->save();
+                $item->set('og_image', $file)->save();
+
+                // Generate, save and set default twitter image/meta.
+                $file = "{$title}-twitter-{$unique}.png";
+                $image = Browsershot::url("{$absolute_url}/social-images/{$id}")
+                    ->windowSize(1200, 600)
+                    ->select('#twitter')
+                    ->save($disk->path($file));
+                $container->makeAsset($file)->save();
+                $item->set('twitter_image', $file)->save();
+            }
         });
 
         Artisan::call('cache:clear');
